requirements.txt
h5py
imageio
matplotlib
numpy
probscale
pygame
pytest
pyyaml==5.4.1
sacred
scipy
seaborn
snakeviz
tensorboard
tensorboard-logger
tensorboardx
torch
torchaudio
torchvision
git+https://github.com/oxwhirl/smac.git
